changes up routing of $surveyId paths, conditionally has a failure splash, gets 95% of the way to displaying the now properly served survey responses, consolidates survey fetching to the root of surveyId path
