+++ server.ts
@@ -3,30 +3,26 @@ import { Question, Survey, SurveyResponse } from "@prisma/client";
 import { NewSurveyPostParams } from "~/routes/surveys.new";
 import prisma from "~/client";
 import cors from "cors";
-import { SurveyResponsePostBody } from "~/routes/surveys.$surveyId";
 const app = express();
 const port = 4000;
 
 app.use(cors());
 app.use(express.json());
 
-app.post(
-  "/answers/submit",
-  async (req: Request<{}, {}, SurveyResponsePostBody>, res: Response) => {
-    const body = req.body;
-    console.log(body);
-    const { answers } = body;
-    const newlyCreatedAnswers = [];
-    for (const newAnswer of answers) {
-      const createdAnswer = await prisma.surveyResponse.create({
-        data: newAnswer
-      });
-      newlyCreatedAnswers.push(createdAnswer);
-    }
-    console.log(newlyCreatedAnswers);
-    res.json(newlyCreatedAnswers);
+app.post("/answers/submit", async (req: Request, res: Response) => {
+  const body = req.body;
+  console.log(body);
+  const { answers } = body;
+  const newlyCreatedAnswers = [];
+  for (const newAnswer of answers) {
+    const createdAnswer = await prisma.surveyResponse.create({
+      data: newAnswer
+    });
+    newlyCreatedAnswers.push(createdAnswer);
   }
-);
+  console.log(newlyCreatedAnswers);
+  res.json(newlyCreatedAnswers);
+});
 
 //create a new survey
 app.post(
@@ -87,7 +83,7 @@ app.get("/surveys/:surveyId/questions", async (req: Request, res: Response) => {
   });
   res.json({ questions });
 });
-//get one survey
+//get one surveyId
 app.get("/surveys/:surveyId", async (req: Request, res: Response) => {
   const survey = await prisma.survey.findUnique({
     where: {
@@ -105,6 +101,25 @@ app.get("/surveys", async (req: Request, res: Response) => {
   const surveys = await prisma.survey.findMany();
 
   res.json({ surveys });
+});
+
+/**
+ * Fetch an entire survey and its questions and answers
+ */
+app.get("/surveys/:surveyId/full", async (req: Request, res: Response) => {
+  const surveyId = req.params.surveyId;
+  //find the survey with surveyId, and bundle all its questions and each question's answers
+  const fetchedSurveyData = await prisma.survey.findUnique({
+    where: { id: surveyId },
+    include: {
+      Questions: {
+        include: {
+          Responses: true
+        }
+      }
+    }
+  });
+  res.json(fetchedSurveyData);
 });
 
 //sends an array of a surveys answers and which questions they go with
